# Fix validation for Haplotypes