--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `Haplotype` validation ([#36](https://github.com/BioJulia/SequenceVariation.jl/issues/36)/[#37](https://github.com/BioJulia/SequenceVariation.jl/pull/37))
+
 ## [0.2.1] - 2023-01-11
 
 ### Added