@MillironX 2023-01-25
> failing test on julia 1.6 seems to be coming [from here](https://github.com/BioJulia/SequenceVariation.jl/pull/37/files#diff-e84e296f931a7cae11eb0b714d8285a9b6897ef391c0a01732c09e6665f7e635R69)

No kidding. Unfortunately I've been kept away from my Julia dev machine since I submitted

> I think it should be as easy as explicitly converting `length(reference(h))` to UInt64

I'm glad it's that simple: I was expected something much more involved. I will test this myself tomorrow morning.