@MillironX 2023-01-28
> Would be nice to not have an x on codecov, but I don't consider that blocking

The irony is that I've added direct tests where there was only indirect coverage before. I think it's "failing" because I've added more lines of code to test, even if every added line is tested.

I'm going to squash some of these commits, add a CHANGELOG entry and merge.